+++ modules/ldapbackend/OBJECTFILES
@@ -1,1 +1,1 @@
-ldapbackend.lo powerldap.lo ldaputils.lo ldapauthenticator.lo
+ldapbackend.lo native.lo master.lo powerldap.lo ldaputils.lo ldapauthenticator.lo
